refactor(frontend-architecture): rearrange the file organization to reflect well the screaming architecture
