+++ frontend/src/modules/auth/core/lib/actions.ts
@@ -1,0 +1,23 @@
+import { authService } from '@/modules/auth/core/services/auth';
+import { defineAction } from 'astro:actions';
+
+export const user = {
+  getUser: defineAction({
+    handler: async (_, context) => {
+      const { data, error } = await authService(context.cookies).getUser();
+      if (error) {
+        throw error;
+      }
+      return data;
+    },
+  }),
+  logout: defineAction({
+    handler: async (_, context) => {
+      const { data, error } = await authService(context.cookies).logout();
+      if (error) {
+        throw error;
+      }
+      return data;
+    },
+  }),
+};
