+++ frontend/src/modules/auth/core/lib/routes.ts
