+++ frontend/src/modules/auth/core/types/oauthProvider.ts
@@ -1,0 +1,9 @@
+import type { OAuthProvidersName } from '@/modules/auth/core/lib/oauthProvider';
+
+export type OAuthProviderName =
+  (typeof OAuthProvidersName)[keyof typeof OAuthProvidersName];
+
+export type OAuthProvider = {
+  id: string;
+  name: OAuthProviderName;
+};
