+++ frontend/src/modules/auth/core/types/role.ts
@@ -1,0 +1,3 @@
+import type { Roles } from '@/modules/auth/core/lib/roles';
+
+export type Role = (typeof Roles)[keyof typeof Roles];
