--- frontend/src/modules/auth/types/oauthProvider.ts
@@ -1,9 +1,0 @@
-import type { OAuthProvidersName } from '@/modules/auth/lib/oauthProvider';
-
-export type OAuthProviderName =
-  (typeof OAuthProvidersName)[keyof typeof OAuthProvidersName];
-
-export type OAuthProvider = {
-  id: string;
-  name: OAuthProviderName;
-};
