--- frontend/src/modules/auth/types/role.ts
@@ -1,3 +1,0 @@
-import type { Roles } from '@/modules/auth/lib/roles';
-
-export type Role = (typeof Roles)[keyof typeof Roles];
