+++ frontend/src/modules/core/layouts/ErrorLayout.astro
@@ -4,9 +4,9 @@ import { type RootLayoutProps } from '@/modules/core/types/layouts';
 import type { ErrorResponse } from '@/modules/core/types/error';
 import { LogoExtended } from '@/modules/core/ui/logoExtended';
 import  H2 from '@/modules/core/ui/H2.astro';
-import { LANDING_ROUTE } from '@/modules/landing/lib/routes';
 import { Link } from '@/modules/core/ui/link';
 import H1 from '@/modules/core/ui/H1.astro';
+import { LANDING_ROUTE } from '@/modules/landing/core/lib/routes';
 
 type Props = RootLayoutProps & ErrorResponse;
