+++ frontend/src/modules/dashboard/instituciones/admin/matriculas/core/layouts/AdminEnrollmentLateralBar.astro
@@ -1,0 +1,47 @@
+---
+import type { AuthRequired, User } from '@/modules/auth/core/types/user';
+import { capitalize } from '@/modules/core/lib/utils';
+import Button from '@/modules/core/ui/button.astro';
+import LogoExtended from '@/modules/core/ui/LogoExtended.astro';
+import { DASHBOARD_ADMIN_INSTITUTION_ENROLLMENT_ROUTE } from '../lib/routes';
+
+// Por ahora solo se requerira un usuario autenticado, ya que el usuario en la base de datos no tiene un rol como campo
+// Puede comportarse como un administrador o un usuario normal, dependiendo de la institucion a la que pertenezca
+type Props = AuthRequired<User>;
+
+const { user } = Astro.props;
+
+const adminOptions = DASHBOARD_ADMIN_INSTITUTION_ENROLLMENT_ROUTE.sub;
+---
+<aside class="flex-initial max-w-2xs px-2 py-4 sticky top-0 h-screen bg-background gap-6 flex flex-col items-center overflow-y-auto">
+  <div>
+    <LogoExtended size="lg"/>
+    <nav class="flex flex-col gap-1">
+      {
+        adminOptions?.map((option) =>
+          <Button
+            as="a"
+            variant="ghost"
+            class="w-full"
+            href={option.fullPath}
+          >{capitalize(option.path || '')}</Button>
+        )
+      }
+    </nav>
+  </div>
+  <div class="min-h-12">
+    <div class="flex items-start gap-3 px-3">
+      <button class="rounded-full aspect-square bg-zinc-300 w-9">
+        <span class="sr-only">Personal photo</span>
+        <img
+          class="w-full h-full object-cover rounded-full"
+          src={user.avatar_url}
+          alt="User profile"
+        />
+      </button>
+      <div class="flex flex-col gap-1">
+        <span class="font-medium">{user.name}</span>
+      </div>
+    </div>
+  </div>
+</aside>
