+++ frontend/src/modules/dashboard/instituciones/admin/matriculas/core/lib/routes.ts
@@ -1,0 +1,11 @@
+import { getRouteFromPath } from '@/modules/core/lib/routes';
+
+// Only for faster matching
+export const DASHBOARD_ADMIN_INSTITUTION_ROUTE = getRouteFromPath(
+  '/dashboard/admin/instituciones/:institution_id',
+);
+
+// Only for faster matching
+export const DASHBOARD_ADMIN_INSTITUTION_ENROLLMENT_ROUTE = getRouteFromPath(
+  '/dashboard/admin/instituciones/:institution_id/matriculas/:matricula_id',
+);
