+++ frontend/src/modules/dashboard/instituciones/estudiante/matriculas/core/componentes/CardHorario.astro
