--- frontend/src/modules/landing/components/DashboardOrSignin.astro
@@ -1,13 +1,0 @@
----
-import SignInModal from '@/modules/auth/components/SignInModal';
-import { authService } from '@/modules/auth/services';
-import Button from '@/modules/core/ui/button.astro';
-import { DASHBOARD_ROUTE } from '@/modules/dashboard/lib/routes';
-
-const { data: user } = await authService(Astro.cookies).getUser();
----
-{user ? (
-<Button as="a" href={DASHBOARD_ROUTE.fullPath}>Panel</Button>
-) : (
-  <SignInModal client:load />
-)}
