--- frontend/src/modules/landing/components/contactUs/ContactUs.astro
@@ -1,12 +1,0 @@
----
-import Section from '@/modules/landing/ui/Section.astro';
-import SectionHeader from '@/modules/landing/ui/SectionHeader.astro';
-import { ContactForm } from '@/modules/landing/components/contactUs/ContactForm';
----
-<Section class="pb-14 ">
-  <SectionHeader>
-    <span slot="title">¿Listo para simplificar tu proceso de matrícula?</span>
-    <span slot="subtitle">Contáctanos y descubre cómo podemos ayudarte</span>
-  </SectionHeader>
-  <ContactForm client:visible />
-</Section>
