--- frontend/src/modules/landing/components/hero/Hero.astro
@@ -1,23 +1,0 @@
----
-import FlyerCounter from '@/modules/landing/components/hero/FlyerCounter';
----
-
-<div class="w-full my-12 grid md:grid-cols-2 grid-rows-[repeat(3,_minmax(350px,_1fr))] md:grid-rows-[repeat(2,_minmax(350px,_1fr))] gap-4">
-  <FlyerCounter
-    client:visible
-    until={20000}
-    exp="peticiones por segundo"
-    className="md:col-span-2"
-  />
-<FlyerCounter
-  client:visible
-  until={1000}
-  exp="matriculados por minuto"
-/>
-<FlyerCounter
-  client:visible
-  until={100}
-  until={50}
-  exp="ms de carga"
-/>
-</div>
