+++ frontend/src/modules/landing/core/layouts/LandingLayout.astro
@@ -1,0 +1,16 @@
+---
+import RootLayout from '@/modules/core/layouts/RootLayout.astro';
+import type { RootLayoutProps } from '@/modules/core/types/layouts';
+import NavBar from '@/modules/landing/core/layouts/NavBar.astro';
+import Footer from '@/modules/landing/core/layouts/NavBar.astro';
+
+type Props = RootLayoutProps;
+const { title } = Astro.props;
+---
+<RootLayout title={title}>
+  <NavBar />
+  <main class="mx-auto max-w-7xl flex flex-col items-center px-2 sm:px-4 md:px-8">
+    <slot />
+  </main>
+  <Footer />
+</RootLayout>
