+++ frontend/src/modules/landing/core/layouts/NavBar.astro
