+++ frontend/src/modules/landing/core/ui/Section.astro
