+++ frontend/src/modules/landing/core/ui/SectionHeader.astro
