--- frontend/src/modules/landing/ui/Label.astro
@@ -1,16 +1,0 @@
----
-import type { HTMLAttributes } from 'astro/types';
-
-type Props = HTMLAttributes<'label'>;
-const { class: className, for: forName, ...rest } = Astro.props;
----
-<label
-  for={forName}
-  class:list={[
-    'font-semibold gap-4',
-    className,
-  ]}
-  {...rest}
->
-  <slot />
-</label>
