--- frontend/src/modules/landing/ui/Required.astro
@@ -1,1 +1,0 @@
-<strong><span class="text-red-600" aria-label="required">*</span></strong>
