--- frontend/src/modules/landing/ui/Textarea.astro
@@ -1,12 +1,0 @@
----
-import { type HTMLAttributes } from 'astro/types';
-type Props = HTMLAttributes<'input'>;
-const { class: className, ...rest } = Astro.props;
----
-<textarea
-  class:list={[
-    'outline-1 outline-slate-400 rounded-md px-2 py-1 focus:outline-blue-600 font-normal',
-    className,
-  ]}
-  {...rest}
->
