+++ frontend/src/pages/dashboard/admin/instituciones/crear/nombre.astro
@@ -3,8 +3,6 @@ import SimplePageLayout from '@/modules/core/layouts/SimplePageLayout.astro';
 import H1 from '@/modules/core/ui/H1.astro';
 import H2 from '@/modules/core/ui/H2.astro';
 import LogoExtended from '@/modules/core/ui/LogoExtended.astro';
-import Button from '@/modules/core/ui/button.astro';
-import Input from '@/modules/landing/ui/Input.astro';
 ---
 <SimplePageLayout title="Dinos el nombrre de tu institución">
   <LogoExtended size="xl" />
@@ -16,10 +14,10 @@ import Input from '@/modules/landing/ui/Input.astro';
       Todos los usuarios de la plataforma verán el nombre que elijas.
     </H2>
   </header>
-  <form class="flex flex-col gap-y-4 w-full max-w-md items-center">
+  <!-- <form class="flex flex-col gap-y-4 w-full max-w-md items-center">
     <Input class="w-full" placeholder="Introduce aquí el nombre" />
     <Button>
       Finalizar
     </Button>
-  </form>
+  </form> -->
 </SimplePageLayout>
