+++ frontend/src/pages/dashboard/index.astro
@@ -1,9 +1,9 @@
 ---
-import { authService } from '@/modules/auth/services';
+import { authService } from '@/modules/auth/core/services/auth';
 import Button from '@/modules/core/ui/button.astro';
-import { LANDING_ROUTE } from '@/modules/landing/lib/routes';
-import { DASHBOARD_ROUTE } from '@/modules/dashboard/lib/routes';
-import DashboardAdminLayout from '@/modules/dashboard/layouts/DashboardAdminLayout.astro';
+import { LANDING_ROUTE } from '@/modules/landing/core/lib/routes';
+import { DASHBOARD_ROUTE } from '@/modules/dashboard/core/lib/routes';
+import AdminEnrollmentLayout from '@/modules/dashboard/instituciones/admin/matriculas/core/layouts/AdminEnrollmentLayout.astro';
 
 const { data: user, error } = await authService(Astro.cookies).getUser();
 
@@ -13,7 +13,7 @@ if (error) {
 
 const route = DASHBOARD_ROUTE;
 ---
-<DashboardAdminLayout title={route.title} user={user}>
+<AdminEnrollmentLayout title={route.title} user={user}>
   <h2>Welcome, {user.name}!</h2>
   <img src={user.avatar_url} alt="xd">
   <p>{user.email}</p>
@@ -21,9 +21,9 @@ const route = DASHBOARD_ROUTE;
     Logout
   </Button>
   <div class="min-h-screen"></div>
-</DashboardAdminLayout>
+</AdminEnrollmentLayout>
 <script>
-import { LANDING_ROUTE } from '@/modules/landing/lib/routes';
+import { LANDING_ROUTE } from '@/modules/landing/core/lib/routes';
 import { actions } from 'astro:actions';
 import { navigate } from 'astro:transitions/client';
