+++ frontend/src/pages/index.astro
@@ -1,8 +1,8 @@
 ---
-import ContactUs from '@/modules/landing/components/contactUs/ContactUs.astro';
-import Hero from '@/modules/landing/components/hero/Hero.astro';
-import LandingLayout from '@/modules/landing/layouts/LandingLayout.astro';
-import { LANDING_ROUTE } from '@/modules/landing/lib/routes';
+import ContactUs from '@/modules/landing/core/components/contactUs/ContactUs.astro';
+import Hero from '@/modules/landing/core/components/hero/Hero.astro';
+import LandingLayout from '@/modules/landing/core/layouts/LandingLayout.astro';
+import { LANDING_ROUTE } from '@/modules/landing/core/lib/routes';
 
 export const prerender = true;
 ---
