+++ frontend/tsconfig.json
@@ -15,6 +15,11 @@
         "./src/*"
       ]
     },
+    "plugins": [
+      {
+        "name": "@astrojs/ts-plugin"
+      }
+    ],
     "verbatimModuleSyntax": true,
     "jsx": "react-jsx",
     "jsxImportSource": "react"
